Address review comments
This is primarily a matter of making more code visible in snippets,
partly by extracting Startup into its own file.
